+++ source/did-wallet-admin/frontend/src/apis/wallet-service-api.ts
@@ -1,0 +1,7 @@
+import { getData } from "../utils/api";
+
+const API_BASE_URL = "/wallet/admin/v1";
+
+export const getWalletServiceInfo = async () => {
+    return getData(API_BASE_URL, "wallet-service/info");
+}