+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [Unreleased] (v1.0.1)
+
+### 🚀 New Features
+- Admin Console
+    - Wallet Service Management
+    - Wallet Management
+    - Admin Management
+- Removed hardcoded logic and sample data from Wallet Server through Admin Console
+  - Removed hardcoded YAML configuration values
+  - Removed sample data previously embedded in Java source code
+  
 ## v1.0.0 (2024-10-18)
 
 ### 🚀 New Features